+++ samples/code-first-ef-core/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+using SemanticTypeModel.EFCore;
+using SemanticTypeModel.Samples.OrderFulfillment.Domain;
+
+[assembly: GenerateSemanticEfModel(typeof(Customer))]